Add functionality to automatically download citygml schema.
This code downloads each schema to a local folder hierarchy.
It also reads the imports and includes and downloads these schemas.
Finally it updates the imports to the local relative path.

This will leave us with a set of local schemas that aren't
referencing any non-local files.
With this the xmlschema library can read it.